requirements.txt
chatterbox-tts
ormsgpack
python-dotenv
requests
torch
torchaudio
